1)  Disabled building of mod-script-pipe on Mac...must build it explicitly. 2)  New HtmlWindow class to add accessibilty to wxHtmlWindow. 3)  Added note and ifdef to start of AudacityApp.cpp about a useful tool to detect memory leaks. 4)  Reworked progress dialog usage...everything should now use the new ProgressDialog class. 5)  Removed Import progress callback in project. 6)  Started work on adding SetName()s to correct what screen readers get for a control label. 7)  Fixed a couple of warnings 8)  Fixed Visual Studio Express issue with solution file. 9)  Fixed incremental link warning on Windows. 10) Converted some files with CRLF line endings.

+++ sf-cvs/trunk/audacity-src/src/effects/nyquist/Nyquist.cpp
@@ -788,7 +788,7 @@ NyquistDialog::NyquistDialog(wxWindow * parent, wxWindowID id,
                wxDefaultPosition, wxSize( 150, -1 ), choices );
          
          int val = ( int )ctrl->val;
-         if( val >= 0 && val < choice->GetCount() )
+         if( val >= 0 && val < (int)choice->GetCount() )
             choice->SetSelection( val );
         
          grid->Add( 10, 10 );
